+++ plugin/evm/atomic/vm/atomic_leaf_handler.go
@@ -1,0 +1,43 @@
+// (c) 2019-2025, Ava Labs, Inc. All rights reserved.
+// See the file LICENSE for licensing terms.
+
+package vm
+
+import (
+	"context"
+	"errors"
+
+	"github.com/ava-labs/avalanchego/codec"
+	"github.com/ava-labs/avalanchego/ids"
+	"github.com/ava-labs/coreth/plugin/evm/message"
+	"github.com/ava-labs/coreth/sync/handlers"
+	"github.com/ava-labs/coreth/sync/handlers/stats"
+	"github.com/ava-labs/libevm/metrics"
+	"github.com/ava-labs/libevm/triedb"
+)
+
+var errUninitialized = errors.New("uninitialized handler")
+
+type uninitializedHandler struct{}
+
+func (h *uninitializedHandler) OnLeafsRequest(ctx context.Context, nodeID ids.NodeID, requestID uint32, leafsRequest message.LeafsRequest) ([]byte, error) {
+	return nil, errUninitialized
+}
+
+// atomicLeafHandler is a wrapper around handlers.LeafRequestHandler that allows for initialization after creation
+type atomicLeafHandler struct {
+	handlers.LeafRequestHandler
+}
+
+// NewAtomicLeafHandler returns a new uninitialzied atomicLeafHandler that can be later initialized
+func NewAtomicLeafHandler() *atomicLeafHandler {
+	return &atomicLeafHandler{
+		LeafRequestHandler: &uninitializedHandler{},
+	}
+}
+
+// Initialize initializes the atomicLeafHandler with the provided atomicTrieDB, trieKeyLength, and networkCodec
+func (a *atomicLeafHandler) Initialize(atomicTrieDB *triedb.Database, trieKeyLength int, networkCodec codec.Manager) {
+	handlerStats := stats.GetOrRegisterHandlerStats(metrics.Enabled)
+	a.LeafRequestHandler = handlers.NewLeafsRequestHandler(atomicTrieDB, trieKeyLength, nil, networkCodec, handlerStats)
+}