+++ plugin/factory/factory.go
@@ -1,0 +1,31 @@
+// (c) 2019-2020, Ava Labs, Inc. All rights reserved.
+// See the file LICENSE for licensing terms.
+
+package factory
+
+import (
+	"github.com/ava-labs/avalanchego/ids"
+	"github.com/ava-labs/avalanchego/snow/engine/snowman/block"
+	"github.com/ava-labs/avalanchego/utils/logging"
+	"github.com/ava-labs/avalanchego/vms"
+
+	"github.com/ava-labs/coreth/plugin/evm"
+	atomicvm "github.com/ava-labs/coreth/plugin/evm/atomic/vm"
+)
+
+var (
+	// ID this VM should be referenced by
+	ID = ids.ID{'e', 'v', 'm'}
+
+	_ vms.Factory = &Factory{}
+)
+
+type Factory struct{}
+
+func (*Factory) New(logging.Logger) (interface{}, error) {
+	return atomicvm.WrapVM(&evm.VM{}), nil
+}
+
+func NewPluginVM() (block.ChainVM, error) {
+	return atomicvm.WrapVM(&evm.VM{IsPlugin: true}), nil
+}