+++ sync/client/mock_client.go
@@ -25,7 +25,7 @@ var (
 // TODO replace with gomock library
 type MockClient struct {
 	codec          codec.Manager
-	leafsHandler   *handlers.LeafsRequestHandler
+	leafsHandler   handlers.LeafRequestHandler
 	leavesReceived int32
 	codesHandler   *handlers.CodeRequestHandler
 	codeReceived   int32
@@ -44,7 +44,7 @@ type MockClient struct {
 
 func NewMockClient(
 	codec codec.Manager,
-	leafHandler *handlers.LeafsRequestHandler,
+	leafHandler handlers.LeafRequestHandler,
 	codesHandler *handlers.CodeRequestHandler,
 	blocksHandler *handlers.BlockRequestHandler,
 ) *MockClient {