+++ RELEASES.md
@@ -7,6 +7,8 @@
 - Implement ACP-176
 - Add `GasTarget` to the chain config to allow modifying the chain's `GasTarget` based on the ACP-176 rules
 
+- Moved atomic codebase from plugin/evm to plugin/evm/atomic package
+
 ## [v0.14.1](https://github.com/ava-labs/coreth/releases/tag/v0.14.1)
 
 - Removed deprecated `ExportKey`, `ExportAVAX`, `Export`, `ImportKey`, `ImportAVAX`, `Import` APIs